IndexedDB/submissions/Opera/Microsoft-port/idbfactory_open4.htm

<!DOCTYPE html>
<html>
<head>  
    <title>IDBFactory.open() - new database has default version</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="../support.js"></script>
    <script>
        var db = null;
        var t = async_test(); 
        
        function RunTest() {
            var rqOpen = window.indexedDB.open(databaseName + "4");
            rqOpen.onupgradeneeded = t.step_func( function(event) {
                db = event.target.result;
                assert_equals(db.version, 1, "db.version");
            });
            rqOpen.onsuccess = t.step_func( function(event) {
                db = event.target.result;
                assert_equals(db.version, 1, "db.version");
                t.done();
            });
            rqOpen.onerror = t.step_func( assert_open_request_error );
        }
        
        add_completion_callback(function() { if(db) db.close(); });
        
        setup(function() {
            var rqDelete = window.indexedDB.deleteDatabase(databaseName);
            rqDelete.onsuccess = t.step_func( RunTest );
        });
    </script>
</head>
<body>
    <div id="log">
    </div>
</body>
</html>
